Create deployment_graph.py
